services/app/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["jalvespinto <jalvespinto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
attrs = "^23.2.0"
cattrs = "^23.2.3"
fastapi = "^0.111.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
python-multipart = "^0.0.9"
Unidecode = "^1.3.8"
asyncpg = "^0.29.0"
aio-pika = "^9.4.0"
yarl = "^1.9.4"
anyio = "^4.0.0"
colorlog = "^6.8.0"
sqlalchemy = "^2.0.25"
apscheduler = "^3.10.4"
pydantic-settings = "^2.1.0"
python-json-logger = "^2.0.7"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"